+++ tests/testthat/test-ggwithinstats.R
@@ -17,7 +17,7 @@ test_that(
         data = data_bugs_2,
         x = condition,
         y = desire,
-        pairwise.comparisons = FALSE,
+        pairwise.display = "none",
         ggsignif.args = list(textsize = 6, tip_length = 0.01),
         point.path.args = list(color = "red"),
         centrality.path.args = list(color = "blue", size = 2, alpha = 0.8),
@@ -33,7 +33,7 @@ test_that(
         data = WRS2::WineTasting,
         x = Wine,
         y = Taste,
-        pairwise.comparisons = FALSE,
+        pairwise.display = "none",
         title = "wine tasting data"
       )
     )
@@ -53,7 +53,7 @@ test_that(
         x = Wine,
         y = Taste,
         results.subtitle = FALSE,
-        pairwise.comparisons = FALSE,
+        pairwise.display = "none",
         ggplot.component = ggplot2::labs(y = "Taste rating")
       )
     )
@@ -68,7 +68,7 @@ test_that(
         centrality.point.args = list(size = 5, alpha = 0.5, color = "darkred"),
         centrality.path = FALSE,
         results.subtitle = FALSE,
-        pairwise.comparisons = FALSE
+        pairwise.display = "none"
       )
     )
   }